requirements.txt
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
rpaframework == 7.1.1
validators == 0.18.2
invoke >= 0.20
rellu >= 0.6
docutils >= 0.14